Move Utils to own file
